Add Monogamous to the Components section
Monogamous is a module that allows you to enforce running you app as a single-instance application. It's not technically limited to usage in Electron, but the examples in the [readme](https://github.com/mnichols/monogamous/blob/master/README.md) are for Electron and there are several issues in Electron were people asked for something like this, so I think it's useful.