Add Electron Talk
This talk was given at Forward 4 Web Summit, Feb 2016 about the "Perks of
the New Desktop".

It begins with an introduction about Electron, a basic overview of how
it works, then goes on to discuss a few key benefits of this development
environment:

1. Native experiences
2. Offline & data storage
3. Modern performance tools
4. Extensibility

The talk was professionally recorded by the Forward Web Summit team and
has both slides and live demos over the course of 45 minutes.