Add a video presented a LNUG: Automatically Build and Publish Node and Electron Applications for Linux
I gave a talk a few weeks ago about automatically building and publishing Node.JS and Electron applications for Linux using Snapcraft. Of particular interest to Electron developers is that `electron-builder` has first class support for snaps and publishing in the snap store gives feature parity with the built-in auto-update features available in Electron for Windows and macOS. Snaps of Electron applications published in the store will be discoverable via the Software store installed by default in all major Linux distributions.

This pull request adds the video of that talk.