Add a link to Conveyor, a packaging and deployment tool
Conveyor is a local CLI tool that creates self-updating packages for every OS from any OS (i.e. it can cross-build and cross-sign packages without CI). It is differentiated from existing tools because it:

1. Can produce fully signed/notarized and self-updating packages for Win/Mac/Linux from any of those operating systems, including Linux CI. It can do this because it doesn't rely on Electron Forge/Builder, nor any native tooling.
2. Getting auto updates doesn't require code changes.
3. It uses Sparkle and MSIX for updates on macOS and Windows, without relying on Squirrel. This makes it easier to deploy to Windows networks.
4. It supports forced update-on-start along with various other nice features.

It's free for open source projects and doesn't rely on any services or SaaS. You can self-host the files it generates, or use S3 or GitHub Releases, for example.